+++ tests/testthat/test-solvect_chlorine.R
@@ -1,8 +1,8 @@
 test_that("solvect_chlorine returns 0's for ct_actual and giardia log when arguments are 0.", {
   water1 <- suppressWarnings(define_water(7.5, 20, 66, toc = 4, uv254 = .2, br = 30))
-  ct1 <- solvect_chlorine(water1, time = 0, residual = 5, baffle = .5)
-  ct2 <- solvect_chlorine(water1, time = 30, residual = 0, baffle = .5)
-  ct3 <- solvect_chlorine(water1, time = 30, residual = 5, baffle = 0)
+  ct1 <- suppressWarnings(solvect_chlorine(water1, time = 0, residual = 5, baffle = .5))
+  ct2 <- suppressWarnings(solvect_chlorine(water1, time = 30, residual = 0, baffle = .5))
+  ct3 <- suppressWarnings(solvect_chlorine(water1, time = 30, residual = 5, baffle = 0))
 
   expect_equal(ct1$ct_actual, 0)
   expect_equal(ct2$glog_removal, 0)
@@ -28,8 +28,8 @@ test_that("solvect_chlorine fails without ph and temp.", {
 test_that("solvect_chlorine correctly uses free_chlorine slot", {
   water1 <- suppressWarnings(define_water(ph = 7.5, temp = 20, toc = 3.5, uv254 = 0.1, br = 50, free_chlorine = 1))
   ct <- solvect_chlorine(water1, time = 30, residual = 5, baffle = 0.3)
-  ct_use <- solvect_chlorine(water1, time = 30, residual = 5, baffle = 0.3, free_cl_slot = "slot_only")
-  ct_use2 <- solvect_chlorine(water1, time = 30, baffle = 0.3, free_cl_slot = "slot_only") # no residual argument
+  ct_use <- suppressWarnings(solvect_chlorine(water1, time = 30, residual = 5, baffle = 0.3, free_cl_slot = "slot_only"))
+  ct_use2 <- suppressWarnings(solvect_chlorine(water1, time = 30, baffle = 0.3, free_cl_slot = "slot_only")) # no residual argument
   ct_use3 <- solvect_chlorine(water1, time = 30, residual = 5, baffle = 0.3, free_cl_slot = "sum_with_residual")
 
 
@@ -53,24 +53,27 @@ test_that("solvect_chlorine works.", {
 
 test_that("solvect_chlorine determines virus log removal", {
   water1 <- suppressWarnings(define_water(ph=7.5, temp=20, toc=3.5, uv254=0.1, br=50))
-  water2 <- suppressWarnings(define_water(ph=10, temp=15, toc=3.5, uv254=0.1, br=50))
-  water3 <- suppressWarnings(define_water(ph=7.5, temp=5, toc=3.5, uv254=0.1, br=50))
-  water4 <- suppressWarnings(define_water(ph=12, temp=20, toc=3.5, uv254=0.1, br=50))
-  water5 <- suppressWarnings(define_water(ph=5, temp=22, toc=3.5, uv254=0.1, br=50))
+  water2 <- suppressWarnings(define_water(ph=7.5, temp=5, toc=3.5, uv254=0.1, br=50))
+  water3 <- suppressWarnings(define_water(ph=7.5, temp=22, toc=3.5, uv254=0.1, br=50))
+  water4 <- suppressWarnings(define_water(ph=9.5, temp=20, toc=3.5, uv254=0.1, br=50))
   
   ct1 <- solvect_chlorine(water1, time = 30, residual = 5, baffle = 0.3)
-  ct2 <- solvect_chlorine(water1, time = 0.5, residual = 1, baffle = 0.3) # ct is too short for virus removal
-  ct3 <- solvect_chlorine(water2, time = 30, residual = 5, baffle = 0.3)
-  ct4 <- solvect_chlorine(water3, time = 10, residual = 2, baffle = 0.3)
-  ct5 <- solvect_chlorine(water4, time = 30, residual = 5, baffle = 0.3)
-  ct6 <- solvect_chlorine(water5, time = 30, residual = 5, baffle = 0.3)
+  ct2 <- solvect_chlorine(water2, time = 10, residual = 2, baffle = 0.3)
   
   expect_equal(ct1$vlog_removal, 4.0)
-  expect_true(is.na(ct2$vlog_removal))
-  expect_equal(ct3$vlog_removal, 4.0)
-  expect_equal(ct4$vlog_removal, 3.0)
-  expect_true(is.na(ct5$vlog_removal))
-  expect_true(is.na(ct6$vlog_removal))
+  expect_equal(ct2$vlog_removal, 3.0)
+  expect_warning(solvect_chlorine(water3, time = 10, residual = 2, baffle = 0.3))
+  expect_warning(solvect_chlorine(water4, time = 10, residual = 2, baffle = 0.3))
+})
+
+test_that("solvect_chlorine warns appropriately about virus log removal", {
+  water1 <- suppressWarnings(define_water(ph=7.5, temp=15, toc=3.5, uv254=0.1, br=50))
+  water2 <- suppressWarnings(define_water(ph=5, temp=20, toc=3.5, uv254=0.1, br=50))
+  water3 <- suppressWarnings(define_water(ph=12, temp=20, toc=3.5, uv254=0.1, br=50))
+  
+  expect_warning(solvect_chlorine(water1, time = 0.5, residual = 1, baffle = 0.3)) # contact time out of range
+  expect_warning(solvect_chlorine(water2, time = 30, residual = 5, baffle = 0.3)) # pH out of range
+  expect_warning(solvect_chlorine(water3, time = 30, residual = 5, baffle = 0.3))
 })
 
 # HELPERS ----